// frontend/src/utils/enums.ts
/* eslint-disable no-unused-vars */
/* eslint-disable import/prefer-default-export */
// eslint-disable-next-line no-shadow
export enum Genres {
  ADVENTURE = 'ADVENTURE',
  FPS = 'FPS',
  MMO = 'MMO',
  MOBILE = 'MOBILE',
  MOBA = 'MOBA',
  PUZZLE = 'PUZZLE',
  RTS = 'RTS',
  RP = 'RP',
  SIMULATION ='SIMULATION',
  SPORTS = 'SPORTS',
}